admin.js orginized, create item added
